+++ common-ux/components/applicationLoadingBar.ts
@@ -9,4 +9,3 @@ import {Component} from 'angular2/core';
 })
 export class ApplicationLoadingBar {
 }
-