+++ common-ux/components/dictionary.ts
@@ -1,0 +1,15 @@
+/* Copyright (c) Microsoft Corporation. All Rights Reserved. */
+/**
+ * Dictionary helper declaration
+ */
+export class Dictionary<T> {
+    /**
+     * Dictionary get and set
+     */
+    [key: string]: T;
+    
+    /**
+     * Convenience constructor
+     */
+    constructor() {}   
+}